split out modules, add language to enum
